validate example STAC Collection/Item against JSON-schema
